Implement missing deconz firmware parameters and fix configuration. Factory reset works now properly

+++ libnymea-zigbee/zigbeeutils.h
@@ -69,7 +69,7 @@ public:
     static QString profileIdToString(const Zigbee::ZigbeeProfile &profileId);
 
     // Generate random data
-    static quint64 generateRandomPanId();
+    static quint16 generateRandomPanId();
 
 
     // Color converter
